Episode1-Chatbot-In-A-Flash/requirements.txt
# requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
google-cloud-aiplatform==1.35.0
python-dotenv==0.19.0
aiofiles==0.8.0